Huge overhaul of API structure and logic.
It should now be much more RESTful, but even more
RESTfulness is coming soon.

+++ api/calendarsapi.py
@@ -13,6 +13,7 @@ import models
 import authutils
 import gapiutils
 import searchutils
+import strings
 
 __author__ = "Alexander Otavka"
 __copyright__ = "Copyright (C) 2015 DHS Developers Club"
@@ -38,28 +39,30 @@ class CalendarsAPI(remote.Service):
         user_service = authutils.get_service(authutils.CALENDAR_API_NAME,
                                              authutils.CALENDAR_API_VERSION)
         app_service = authutils.get_service(
-            authutils.CALENDAR_API_NAME,
-            authutils.CALENDAR_API_VERSION,
-            AppAssertionCredentials(authutils.SERVICE_ACCOUNT_SCOPES)
-        )
+                authutils.CALENDAR_API_NAME,
+                authutils.CALENDAR_API_VERSION,
+                AppAssertionCredentials(authutils.SERVICE_ACCOUNT_SCOPES))
         all_calendars = (gapiutils.get_calendars(user_service) +
                          gapiutils.get_calendars(app_service))
 
         # Filter out calendars not added in user's ndb
         user_key = models.get_user_key(user_id)
-        chosen_ndb = models.Calendar.query(ancestor=user_key).fetch()
+        chosen_ndb = models.Calendar.query(ancestor=user_key)
+        if request.hidden is not None:
+            chosen_ndb.filter(models.Calendar.hidden == request.hidden)
         chosen_calendars = []
-        for entity in chosen_ndb:
+        for entity in chosen_ndb.fetch():
             for cal in all_calendars:
-                if cal.calendar_id == entity.key.string_id():
+                if cal.calendarId == entity.key.string_id():
                     cal.hidden = entity.hidden
+                    if cal.hidden is None:
+                        cal.hidden = False
                     chosen_calendars.append(cal)
                     break
             else:
-                logging.info(
-                    "Deleted: unbound Calendar entity with calendar_id = " +
-                    "\"{}\" and user_id = \"{}\"."
-                    .format(entity.key.string_id(), user_id))
+                logging.info(strings.LOGGING_DELETE_UNBOUND_CALENDAR
+                             .format(calendar_id=entity.key.string_id(),
+                                     user_id=user_id))
                 entity.key.delete()
 
         # Sort and search
@@ -72,57 +75,61 @@ class CalendarsAPI(remote.Service):
 
         return messages.CalendarCollection(items=chosen_calendars)
 
-    @endpoints.method(messages.CALENDAR_ID_RESOURCE, message_types.VoidMessage,
-                      http_method="POST", path="/calendars")
-    def insert(self, request):
+    @staticmethod
+    def get_calendar_entity(calendar_id):
         """
-        Add a calendar to the user's list.
+        Retrieve or create a calendar entity from a calendar id.
 
-        :type request: messages.CALENDAR_ID_RESOURCE
+        :type calendar_id: str
+        :rtype: models.Calendar
         """
         user_id = authutils.require_user_id()
 
-        cal_id = request.calendar_id
+        # Get the ndb entity
         user_key = models.get_user_key(user_id)
-        model = models.Calendar(id=cal_id, parent=user_key)
-        model.put()
-        return message_types.VoidMessage()
+        entity = ndb.Key(models.Calendar, calendar_id,
+                         parent=user_key).get()
+        if entity is None:
+            entity = models.Calendar(id=calendar_id, parent=user_key)
+        return entity
 
-    @endpoints.method(messages.CALENDAR_PATCH_RESOURCE,
-                      messages.CalendarProperties,
-                      http_method="PATCH", path="{calendar_id}")
+    @endpoints.method(messages.CALENDAR_WRITE_RESOURCE,
+                      messages.CalendarWriteProperties,
+                      http_method="PATCH", path="{calendarId}")
     def patch(self, request):
         """
         Update a calendar's data.
 
-        Only Calendar.hidden can be changed.
-
-        :type request: messages.CALENDAR_PATCH_RESOURCE
+        :type request: messages.CALENDAR_WRITE_RESOURCE
         """
-        user_id = authutils.require_user_id()
+        entity = self.get_calendar_entity(request.calendarId)
 
-        # Get the ndb model
-        cal_id = request.calendar_id
-        user_key = models.get_user_key(user_id)
-        model = ndb.Key(models.Calendar, cal_id, parent=user_key).get()
-        if model is None:
-            raise endpoints.NotFoundException(
-                "No calendar with id of \"{}\" in user's list.".format(cal_id))
+        # Set properties from request on the entity
+        if request.hidden is not None:
+            entity.hidden = request.hidden
 
-        # Set properties from request on the model
-        hidden = request.hidden
-        if hidden is not None:
-            model.hidden = hidden
+        entity.put()
+        return messages.CalendarWriteProperties(hidden=entity.hidden)
 
-        model.put()
-        return messages.CalendarProperties(
-            calendar_id=cal_id,
-            hidden=model.hidden,
-        )
+    @endpoints.method(messages.CALENDAR_WRITE_RESOURCE,
+                      messages.CalendarWriteProperties,
+                      http_method="PUT", path="{calendarId}")
+    def put(self, request):
+        """
+        Set a calendar's data, or create a calendar from data.
 
-    @endpoints.method(messages.CALENDAR_ID_RESOURCE,
-                      messages.CalendarProperties,
-                      http_method="DELETE", path="{calendar_id}")
+        :type request: messages.CALENDAR_WRITE_RESOURCE
+        """
+        entity = CalendarsAPI.get_calendar_entity(request.calendarId)
+
+        # Set properties from request on the entity
+        entity.hidden = request.hidden
+
+        entity.put()
+        return messages.CalendarWriteProperties(hidden=entity.hidden)
+
+    @endpoints.method(messages.CALENDAR_ID_RESOURCE, message_types.VoidMessage,
+                      http_method="DELETE", path="{calendarId}")
     def delete(self, request):
         """
         Remove a calendar from a user's list.
@@ -131,12 +138,17 @@ class CalendarsAPI(remote.Service):
         """
         user_id = authutils.require_user_id()
 
-        cal_id = request.calendar_id
         user_key = models.get_user_key(user_id)
-        key = ndb.Key(models.Calendar, cal_id, parent=user_key)
-        entity = key.get()
+        key = ndb.Key(models.Calendar, request.calendarId, parent=user_key)
+        if key.get() is None:
+            raise endpoints.NotFoundException(
+                    strings.ERROR_CALENDAR_NOT_FOUND.format(
+                            calendar_id=request.calendarId))
         key.delete()
-        return messages.CalendarProperties(
-            calendar_id=cal_id,
-            hidden=entity.hidden,
-        )
+
+        # Does not delete the calendar's events' data, just in case the user
+        # wants to undo, they can re-add the calendar, and have all of their
+        # starred and hidden events remain.  I'm still not sure if that's the
+        # best behavior, but I'm going to err on the side of least destruction.
+
+        return message_types.VoidMessage()
